pyproject.toml
[tool.poetry]
name = "us-electonics"
version = "0.1.0"
description = ""
authors = ["Abdulrahman Dawoud <dawoud.abdulrahman.2004@gmail.com>"]
readme = "README.md"
packages = [{include = "us_electonics"}]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
django = "^4.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
